Split include formats into own classes
